feat: single sign-model and UI component
