+++ src/renderer/features/operations/OperationSign/index.ts
@@ -1,1 +1,3 @@
 export { OperationSign } from './ui/OperationSign';
+export { SigningSwitch } from './ui/SigningSwitch';
+export { signModel } from './model/sign-model';
