+++ src/renderer/features/operations/OperationSign/model/sign-model.ts
