+++ src/renderer/features/operations/index.ts
@@ -1,3 +1,3 @@
 export { OperationFooter, OperationHeader } from './init';
-export { OperationSign } from './OperationSign';
+export { OperationSign, SigningSwitch, signModel } from './OperationSign';
 export { OperationsFilter } from './OperationsFilter';
