+++ src/renderer/pages/Operations/components/modals/ApproveTx.tsx
@@ -15,7 +15,7 @@ import { Submit } from '../ActionSteps/Submit';
 import { Confirmation } from '../ActionSteps/Confirmation';
 import { SignatorySelectModal } from './SignatorySelectModal';
 import { useMultisigEvent } from '@entities/multisig';
-import { OperationSign } from '@features/operations';
+import { SigningSwitch } from '@features/operations';
 import { permissionUtils, walletModel } from '@entities/wallet';
 import { priceProviderModel } from '@entities/price';
 import type { Address, HexString, Timepoint, MultisigAccount, Account } from '@shared/core';
@@ -248,7 +248,7 @@ const ApproveTx = ({ tx, account, connection }: Props) => {
         )}
 
         {activeStep === Step.SIGNING && approveTx && connection.api && signAccount && (
-          <OperationSign
+          <SigningSwitch
             chainId={tx.chainId}
             api={connection.api}
             addressPrefix={connection?.addressPrefix}
