+++ src/renderer/widgets/AddProxyModal/ui/AddProxy.tsx
@@ -5,10 +5,10 @@ import { useModalClose } from '@shared/lib/hooks';
 import { OperationTitle } from '@entities/chain';
 import { useI18n } from '@app/providers';
 import type { Chain } from '@shared/core';
+import { OperationSign } from '@features/operations';
 import { Step } from '../lib/types';
 import { AddProxyForm } from './AddProxyForm';
 import { Confirmation } from './Confirmation';
-import { SignProxy } from './SignProxy';
 import { SubmitProxy } from './SubmitProxy';
 import { addProxyUtils } from '../lib/add-proxy-utils';
 import { addProxyModel } from '../model/add-proxy-model';
@@ -35,7 +35,7 @@ export const AddProxy = () => {
       {addProxyUtils.isConfirmStep(step) && (
         <Confirmation onGoBack={() => addProxyModel.events.stepChanged(Step.INIT)} />
       )}
-      {addProxyUtils.isSignStep(step) && <SignProxy onGoBack={() => addProxyModel.events.stepChanged(Step.CONFIRM)} />}
+      {addProxyUtils.isSignStep(step) && <OperationSign onGoBack={() => addProxyModel.events.stepChanged(Step.CONFIRM)} />}
     </BaseModal>
   );
 };
