--- src/renderer/widgets/AddProxyModal/ui/SignProxy.tsx
@@ -1,39 +1,0 @@
-import { useUnit } from 'effector-react';
-import { UnsignedTransaction } from '@substrate/txwrapper-polkadot';
-
-import type { HexString } from '@shared/core';
-import { OperationSign } from '@features/operations';
-import { signModel } from '../model/sign-model';
-
-type Props = {
-  onGoBack: () => void;
-};
-
-export const SignProxy = ({ onGoBack }: Props) => {
-  const api = useUnit(signModel.$api);
-  const signStore = useUnit(signModel.$signStore);
-  const signerWallet = useUnit(signModel.$signerWallet);
-
-  if (!api || !signStore || !signerWallet) return null;
-
-  const onSignResult = (signatures: HexString[], unsignedTxs: UnsignedTransaction[]) => {
-    signModel.events.dataReceived({
-      signature: signatures[0],
-      unsignedTx: unsignedTxs[0],
-    });
-  };
-
-  return (
-    <OperationSign
-      api={api}
-      chainId={signStore.chain.chainId}
-      addressPrefix={signStore.chain.addressPrefix}
-      signerWaller={signerWallet}
-      accounts={[signStore.account]}
-      signatory={signStore.signatory || undefined}
-      transactions={[signStore.transaction]}
-      onGoBack={onGoBack}
-      onResult={onSignResult}
-    />
-  );
-};
