+++ src/renderer/widgets/AddPureProxiedModal/model/add-pure-proxied-model.ts
@@ -7,26 +7,18 @@ import { Transaction, TransactionType, TxWrapper, transactionService, WrapperKin
 import { toAddress } from '@shared/lib/utils';
 import { walletSelectModel } from '@features/wallets';
 import { walletModel, walletUtils } from '@entities/wallet';
-import {
-  type MultisigAccount,
-  type ProxyGroup,
-  type NoID,
-  ProxyType,
-  AccountId,
-  PartialProxiedAccount,
-  ProxyVariant,
-  Account,
-} from '@shared/core';
+import type { MultisigAccount, ProxyGroup, NoID, AccountId, PartialProxiedAccount, Account } from '@shared/core';
+import { ProxyType, ProxyVariant } from '@shared/core';
 import { proxyModel, proxyUtils } from '@entities/proxy';
 import { networkModel } from '@entities/network';
 import { balanceSubModel } from '@features/balances';
+import { proxiesModel } from '@features/proxies';
+import { signModel } from '@features/operations';
 import { Step, TxWrappers, AddPureProxiedStore } from '../lib/types';
 import { formModel } from './form-model';
 import { confirmModel } from './confirm-model';
-import { signModel } from './sign-model';
 import { submitModel } from './submit-model';
 import { subscriptionService } from '@entities/chain';
-import { proxiesModel } from '@features/proxies';
 
 const stepChanged = createEvent<Step>();
