+++ src/renderer/widgets/Transfer/model/transfer-model.ts
@@ -3,10 +3,10 @@ import { spread, delay } from 'patronum';
 import { NavigateFunction } from 'react-router-dom';
 
 import { Transaction } from '@entities/transaction';
+import { signModel } from '@features/operations';
 import { Step, TransferStore, NetworkStore } from '../lib/types';
 import { formModel } from './form-model';
 import { confirmModel } from './confirm-model';
-import { signModel } from './sign-model';
 import { submitModel } from './submit-model';
 import { Paths } from '@shared/routes';
